Fix print statements and runtime
Print statements were changed to reflex Python 3.x syntax and runtime was increased from 3600 to 4200 seconds. 

Run time was increased to allow for 12 images to be taken throughout the duration of the experiment. Accounts for time delay attributed to communication between computer and pieces of equipment,. Will run for slightly longer than 1 hour now. 

Experiment was run on 4/3/20 and code was fully functional.